Use DeGAUSS Container Template

Creates all the necessary files to create a DeGAUSS container. The container/geomarker name is assumed to be the basename of the working directory and the version of R and renv is taken from the calling environment. This function calls all of the individual dht::use_degauss_*() functions to create the following:

Dockerfile

Makefile

README.md

entrypoint.R

.dockerignore

test/my_address_file_geocoded.csv

LICENSE.md GPL license

.github/workflows/build-deply.yaml

Usage

use_degauss_container(geomarker = getwd(), version = "0.1.0", ...)

use_degauss_dockerfile(geomarker = getwd(), version, ...)

use_degauss_makefile(geomarker = getwd(), ...)

use_degauss_readme(geomarker = getwd(), version = "0.1.0", ...)

use_degauss_githook_readme_rmd(geomarker = getwd(), ...)

use_degauss_entrypoint(geomarker = getwd(), version = "0.1.0", ...)

use_degauss_dockerignore(geomarker = getwd(), ...)

use_degauss_tests(geomarker = getwd(), ...)

use_degauss_license(geomarker = getwd(), ...)

use_degauss_github_actions(geomarker = getwd(), ...)

Arguments

geomarker

path to folder where DeGAUSS container files are to be added;

version

string of version number used in freshly created README and entrypoint.R; defaults to "0.1.0"

...

arguments passed to render_degauss_template (overwrite) defaults to the current working directory